@@ -13,46 +13,62 @@
   "private": true,
   "dependencies": {
     "@abacritt/angularx-social-login": "^1.2.5",
-    "@angular/animations": "^15.2.10",
+    "@angular/animations": "^16.2.12",
     "@angular/cdk": "16.0.2",
-    "@angular/common": "^15.2.10",
-    "@angular/compiler": "^15.2.10",
-    "@angular/core": "^15.2.10",
-    "@angular/forms": "^15.2.10",
+    "@angular/common": "^16.2.12",
+    "@angular/compiler": "^16.2.12",
+    "@angular/core": "^16.2.12",
+    "@angular/forms": "^16.2.12",
     "@angular/material": "16.0.2",
     "@angular/material-luxon-adapter": "16.0.2",
-    "@angular/platform-browser": "^15.2.10",
-    "@angular/platform-browser-dynamic": "^15.2.10",
+    "@angular/platform-browser": "^16.2.12",
+    "@angular/platform-browser-dynamic": "^16.2.12",
+    "crypto-js": "3.3.0",
     "@angular/pwa": "^0.5.3",
-    "@angular/router": "^15.2.10",
+    "@angular/router": "^16.2.12",
     "@popperjs/core": "^2.11.6",
     "@types/quill": "^1.3.10",
-    "angular-google-charts": "^2.2.3",
+    "apexcharts": "3.40.0",
     "bootstrap": "^5.2.3",
     "jquery": "^3.6.4",
     "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
+    "highlight.js": "11.8.0",
+    "luxon": "^3.5.0",
+    "ng-apexcharts": "1.7.6",
     "ngx-bootstrap": "9.0.0",
     "ngx-bootstrap-icons": "^1.9.2",
     "ngx-quill": "^17.0.0",
+    "perfect-scrollbar": "1.5.5",
     "quill": "^1.3.7",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.11",
-    "@angular/cli": "~15.2.11",
-    "@angular/compiler-cli": "^15.2.10",
+    "@angular-devkit/build-angular": "^16.2.16",
+    "@angular/cli": "~16.2.16",
+    "@angular/compiler-cli": "^16.2.12",
     "@types/d3": "^7.4.0",
     "@types/d3-shape": "^3.1.1",
     "@types/jasmine": "~4.0.0",
     "@types/lodash": "^4.14.191",
+    "@types/chroma-js": "2.4.0",
+    "@types/lodash-es": "4.17.7",
+    "@tailwindcss/typography": "0.5.9",
+    "@types/highlight.js": "10.1.0",
+    "@types/luxon": "3.3.0",
+    "@types/crypto-js": "3.1.47",
     "jasmine-core": "~4.3.0",
+    "chroma-js": "2.4.2",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
+    "lodash": "4.17.21",
+    "postcss": "8.4.24",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~4.9.5"
+    "tailwindcss": "3.3.2",
+     "typescript": "5.0.4"
   }
 }
