@@ -28,14 +28,14 @@ export class FuseVerticalNavigationAsideItemComponent implements OnChanges, OnIn
     static ngAcceptInputType_skipChildren: BooleanInput;
     /* eslint-enable @typescript-eslint/naming-convention */
 
-    @Input() activeItemId: string;
-    @Input() autoCollapse: boolean;
-    @Input() item: FuseNavigationItem;
-    @Input() name: string;
-    @Input() skipChildren: boolean;
+    @Input() activeItemId!: string;
+    @Input() autoCollapse!: boolean;
+    @Input() item!: FuseNavigationItem;
+    @Input() name!: string;
+    @Input() skipChildren!: boolean;
 
     active: boolean = false;
-    private _fuseVerticalNavigationComponent: FuseVerticalNavigationComponent;
+    private _fuseVerticalNavigationComponent!: FuseVerticalNavigationComponent;
     private _unsubscribeAll: Subject<any> = new Subject<any>();
 
     /**
