@@ -28,13 +28,13 @@ export class FuseVerticalNavigationCollapsableItemComponent implements OnInit, O
     static ngAcceptInputType_autoCollapse: BooleanInput;
     /* eslint-enable @typescript-eslint/naming-convention */
 
-    @Input() autoCollapse: boolean;
-    @Input() item: FuseNavigationItem;
-    @Input() name: string;
+    @Input() autoCollapse!: boolean;
+    @Input() item!: FuseNavigationItem;
+    @Input() name!: string;
 
     isCollapsed: boolean = true;
     isExpanded: boolean = false;
-    private _fuseVerticalNavigationComponent: FuseVerticalNavigationComponent;
+    private _fuseVerticalNavigationComponent!: FuseVerticalNavigationComponent;
     private _unsubscribeAll: Subject<any> = new Subject<any>();
 
     /**
