@@ -41,7 +41,7 @@ export class FuseLoadingService
     /**
      * Getter for progress
      */
-    get progress$(): Observable<number>
+    get progress$(): Observable<number | null>
     {
         return this._progress$.asObservable();
     }
