@@ -1,8 +1,8 @@
 import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { Router } from "@angular/router";
-import { PreferencesService } from "src/app/shared/services/preferences.service";
-import { API_URL } from "src/environments/environment";
+import { PreferencesService } from "app/shared/services/preferences.service";
+import { API_URL } from "environments/environment";
 
 @Injectable({
     providedIn: 'root'
