@@ -4,7 +4,7 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { PreviousExamsRoutingModule } from './previous-exams-routing.module';
 import { PreviousExamsComponent } from './pages/previous-exams/previous-exams.component';
 import { HttpClientModule } from '@angular/common/http';
-import { SharedModule } from 'src/app/shared/shared.module';
+import { SharedModule } from 'app/shared/shared.module';
 
 
 
